Implemented joints
So from what I have learned on joints so far spring joints attached to a rigidbody is our best bet. I am having some issues with the code still but am trying to iron out how to get the mouse to move the hammer without the character tweaking or flying.